% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_connection.R
\name{checkConnection}
\alias{checkConnection}
\title{checkConnection}
\usage{
checkConnection(username, password, host = "lyra.qut.edu.au", port = 22)
}
\arguments{
\item{username}{Username for your QUT HPC account}

\item{password}{PAssword for your QUT HPC account}

\item{host}{Address to QUT HPC}

\item{port}{Port to use for SSH. (Default = 22).}
}
\value{
Just some fill in for now
}
\description{
Just some fill-in for now
}
\details{
Just some fill-in for now
}

