% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_Packages.R
\name{getDependencies}
\alias{getDependencies}
\title{Get package source files and install to a personal library on the HPC file
server}
\usage{
getDependencies(credentials, packages.toinstall, local.dest, remote.dest,
  host = "lyra.qut.edu.au", port = 22)
}
\arguments{
\item{credentials}{Vector of username and password (<Username>, <Password>).
REQUIRED}

\item{local.dest}{Path to store the source files locally before moving to the
HPC file server.}

\item{remote.dest}{Path on the remote device for the Personal Library to install too.}

\item{host}{Address for the host device. (Default="lyra.qut.edu.au", QUT's HPC).}

\item{port}{Port to use for SSH. (Default = 22).}

\item{packages.install}{Vector of packages to be installed. REQUIRED}

\item{remote.library}{Path to personal library on the HPC file server.}
}
\value{

}
\description{
Automates a recursive search for package source files and their dependencies.
Using dependency information a specific install order is  prepared to
compensate for dependency trees. It will also attempt to identify when a
particular module is needed for installation. Once prepared the files are
copied to the HPC file server and installed to a personal library for future
use.
}
\details{

}

