% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logging_Environment.R
\docType{data}
\name{remoteLyraR.Env}
\alias{remoteLyraR.Env}
\title{logging}
\format{An object of class \code{environment} of length 0.}
\usage{
remoteLyraR.Env
}
\description{
logging
}
\keyword{datasets}

