% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitCommandToLyra.R
\name{submitCommandToLyra}
\alias{submitCommandToLyra}
\title{Submit command to HPC}
\usage{
submitCommandToLyra(command, username, password, host = "lyra.qut.edu.au",
  port = 22)
}
\arguments{
\item{command}{TO BE ADDED}

\item{username}{TO BE ADDED}

\item{password}{TO BE ADDED}

\item{host}{Host name for the remote server (default: lyra.qut.edu.au)}

\item{port}{Port number to be used for SSH and SCP to the host (defualt: 22)}
}
\description{
Submit command to HPC
}

