% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitCommandToLyraWindows.R
\name{submitCommandToLyra.Windows}
\alias{submitCommandToLyra.Windows}
\title{Submit command to HPC from Windows.}
\usage{
submitCommandToLyra.Windows(command, username, password,
  host = "lyra.qut.edu.au", port = 22)
}
\description{
Submit command to HPC from Windows.
}

