% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_locale.R
\name{tree}
\alias{tree}
\title{Submit Job to HPC (Remotely Stored Project)}
\usage{
tree(directory)
}
\arguments{
\item{directory}{The Path to the folder to search recursively}
}
\description{
Submit Job to HPC (Remotely Stored Project)
}

