% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFolderFile.R
\name{checkRemoteFolderFileExists}
\alias{checkRemoteFolderFileExists}
\title{checkFolderExists}
\usage{
checkRemoteFolderFileExists(directory, file, username, password,
  host = "lyra.qut.edu.au", port = 22)
}
\description{
checkFolderExists
}

