% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitCommandToLyraWindows.R
\name{submitCommandToLyra.Windows}
\alias{submitCommandToLyra.Windows}
\title{Submit command to HPC from Windows.}
\usage{
submitCommandToLyra.Windows(command, username, password,
  host = "lyra.qut.edu.au", port = 22)
}
\arguments{
\item{host}{Host name for the remote server (default: lyra.qut.edu.au)}

\item{port}{Port number to be used for SSH and SCP to the host (defualt: 22)}
}
\description{
Submit command to HPC from Windows.
}

