% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_remote.R
\name{submitRemote}
\alias{submitRemote}
\title{Submit Job to HPC (Remotely Stored Project)}
\usage{
submitRemote(credentials, remote.folder, script.file, submission.file, data,
  quiet = FALSE, host = "lyra.qut.edu.au", port = 22,
  ignore.warning = FALSE)
}
\arguments{
\item{credentials}{Vector of username and password (<Username>, <Password>). REQUIRED}

\item{remote.folder}{Root directory for the project stored on the remote file server.REQUIRED}

\item{script.file}{R file to be called. REQUIRED}

\item{submission.file}{ADD SOON}

\item{data}{ADD SOON. REQUIRED}

\item{quiet}{Turns on quiet mode, disabling all messages except warnings and errors (default: FALSE)}

\item{host}{Host name for the remote server (default: lyra.qut.edu.au)}

\item{port}{Port number to be used for SSH and SCP to the host (defualt: 22)}

\item{ignore.warning}{Suppress warning messages (default: FALSE)}
}
\value{
A structure (dataframe or datatable, matches the structure of \code{data}) which contains the details of every
submission. Each row of structure contains all the job details (job name, memory, wall time, ncpus, jobid) as well as
the arguments submitted (both user defined and provided).

The structure is also saved as a .rds file as well as parsed into a DSV file (space-delimited). These files are saved
to the \code{remote.folder}.
}
\description{
Submit Job to HPC where project files are already stored on the file server.
}
\details{
Keep in mind that four parameters, credentials, remote.folder, script.file, and data are required.The
remote fileserver must contain the directory \code{remote.folder} and \code{script.file} must be found
in the top level of \code{remote.folder}.

Before attempting to submit the job the \code{credentials} and existence of \code{remote.folder} and
\code{script.file} are checked. An error is given if any fail.

\code{data} must be either a data.frame or a data.table which have the column names:
\itemize{
 \item{"JOBNAME"}{ Name of the job to be submitted to HPC. If \code{REPEAT} != 1 the job name will
 be affixed an identifier. While not necessary it is recommended each job name in the submission stack
 is unqiue (A warning will be produced if a conflixt exists). Only letters, numbers and underscores can
 be used, job name must begin with a letter (Violation to this will produce an error). \emph{REQUIRED}.}
 \item{"MEMORY"}{ The memory required for the job. Example: "100mb" or "2gb" requests 100 megabytes or
 2 gigabytes respectively. \emph{REQUIRED}.}
 \item{"WALLTIME"}{ The time requested for the job on the HPC in the form "<hours>:<mins>:<secs>".
 Example: "1:30:00" will request 1hour and 30minutes. A single number will be converted to hours and
 minutes, for example "1.5" will converted to "1:30:00". \emph{REQUIRED}.}
 \item{"NCPUS"}{ Number of CPUs to be requested from the node for parallelisation. \emph{NOT REQUIRED, DEFAULT: 1}.}
 \item{"DONOTRUNJOB"}{ If \code{TRUE} the job is not submitted. Useful for submitting larger structures
 of parameters without partitioning them into smaller chunks. \emph{NOT REQUIRED, DEFAULT: FALSE}.}
 \item{"REPEAT"}{ Resubmit a particular job \code{REPEAT} times. Useful when submitting a simulation that includes
 random elements. Owing to how the Random Number Generator seeds, each job could initialise with the same seed.
 Due to this a random integer is provided to each submission to mitigate this issue. \emph{NOT REQUIRED, DEFAULT: 1}.}
}

In addition to these columns are any arguments to be passed to the user script in the form \code{<col_name>=<value>}.
}

